refactor(code_gen): extract http builtins into builtins/http.rs
